AuthenticationGuard and RoleGuard implemented

+++ src/app/controllers/user.controller.ts
@@ -1,20 +1,17 @@
 import { Body, Controller, Delete, Get, Param, Post, Put, UseGuards } from '@nestjs/common';
 import { UserService } from '../services/user.service';
-import UserInterface from '../interfaces/user.interface';
 import { CreateUserDTO, UpdateUserDTO } from '../schema/user.schema';
 import UserEntity from '../db/entities/user.entity';
-import { AccessService } from '../services/access.service';
 import EAccess from '../enums/access.enum';
 import ServiceResponse from '../services/ServiceResponse';
-import AuthService from '../services/auth.service';
 import AuthenticationGuard from '../guards/authentication.guard';
 import RolesGuard from '../guards/roles.guard';
+import { GetUser } from '../utils/getUser.decorator';
+import EMessages from '../enums/EMessages';
 
 @Controller('user')
 export default class UserController {
-  constructor(private readonly userService: UserService,
-              private readonly authService: AuthService,
-              private readonly accessService: AccessService) {
+  constructor(private readonly userService: UserService) {
   }
 
   @Get()
@@ -22,10 +19,10 @@ export default class UserController {
   async findAll(): Promise<ServiceResponse> {
     // if (await this.accessService.getAccess() === EAccess.MANAGER || await this.accessService.getAccess() === EAccess.ADMIN) {
       const user: UserEntity[] = await this.userService.findAll();
-      if (!user) {
+      if (user.length === 0) {
         return ServiceResponse.error('no users found');
       } else {
-        return ServiceResponse.success(user);
+        return ServiceResponse.success(user, EMessages.RESOURCE_FOUND);
       }
     // } else {
     //   return ServiceResponse.error('You don\'t have access to view the records, Please log In as MANAGER or ADMIN');
@@ -36,7 +33,7 @@ export default class UserController {
   @UseGuards(AuthenticationGuard, new RolesGuard([EAccess.MANAGER, EAccess.ADMIN]))
   async findById(@Param('userName')userName: string): Promise<ServiceResponse> {
     // if (await this.accessService.getAccess() === EAccess.MANAGER || await this.accessService.getAccess() === EAccess.ADMIN) {
-      const user: UserInterface = await this.userService.findByUserName(userName);
+      const user: UserEntity = await this.userService.findByUserName(userName);
       if (!user) {
         return ServiceResponse.error(`user \"${userName}\" not found`);
       } else {
@@ -49,13 +46,18 @@ export default class UserController {
 
   @Post('/new')
   @UseGuards(AuthenticationGuard, new RolesGuard([EAccess.MANAGER, EAccess.ADMIN]))
-  async createUser(@Body() createUserDTO: CreateUserDTO): Promise<ServiceResponse> {
+  async createUser(@Body() createUserDTO: CreateUserDTO, @GetUser() thisUser: UserEntity): Promise<ServiceResponse> {
     // if (await this.accessService.getAccess() === EAccess.MANAGER || await this.accessService.getAccess() === EAccess.ADMIN) {
-      const user: UserInterface = await this.userService.findByUserName(createUserDTO.userName);
+      const user: UserEntity = await this.userService.findByUserName(createUserDTO.userName);
       if (!user) {
-        if (!createUserDTO.access || !createUserDTO.password) {
-          return ServiceResponse.error('details missing, Please try again');
+        if (!createUserDTO.password) {
+          return ServiceResponse.error('Enter Password & Please try again');
         } else {
+          if (thisUser.access === EAccess.MANAGER) {
+            createUserDTO.access = EAccess.USER;
+          } else if (!createUserDTO.access) {
+            createUserDTO.access = EAccess.USER;
+          }
           return ServiceResponse.success(await this.userService.createUser(createUserDTO));
         }
       } else {
@@ -68,22 +70,26 @@ export default class UserController {
 
   @Delete('/remove/:userName')
   @UseGuards(AuthenticationGuard, new RolesGuard([EAccess.MANAGER, EAccess.ADMIN]))
-  async removeUser(@Param('userName')userName: string): Promise<ServiceResponse> {
+  async removeUser(@Param('userName')userName: string, @GetUser() thisUser: UserEntity): Promise<ServiceResponse> {
     // if (await this.accessService.getAccess() === EAccess.MANAGER || await this.accessService.getAccess() === EAccess.ADMIN) {
-      const user: UserInterface = await this.userService.findByUserName(userName);
+      const user: UserEntity = await this.userService.findByUserName(userName);
       if (!user) {
         return ServiceResponse.error(`user \"${userName}\" not found`);
-      } else {
-        return ServiceResponse.success(await this.userService.removeUser(userName));
       }
-    // } else {
-    //   return ServiceResponse.error('You don\'t have access to create a record, Please log In as MANAGER or ADMIN');
-    // }
+      if ( (thisUser.access === EAccess.MANAGER && user.access === EAccess.USER) || thisUser.access === EAccess.ADMIN ) {
+        return ServiceResponse.success(await this.userService.removeUser(userName));
+      } else {
+      return ServiceResponse.error(EMessages.UNAUTHORIZED_REQUEST);
+    }
   }
 
   @Put('/update')
-  async updateUser(@Body() updateUserDTO: UpdateUserDTO): Promise<ServiceResponse> {
-    if (await this.accessService.getAccess() === EAccess.USER ) {
+  @UseGuards(AuthenticationGuard)
+  async updateUser(@Body() updateUserDTO: UpdateUserDTO, @GetUser() thisUser: UserEntity): Promise<ServiceResponse> {
+    if (!thisUser) {
+      return ServiceResponse.error('Please LOG IN to perform an action');
+    }
+    if ( thisUser.access === EAccess.USER ) {
       if ( !updateUserDTO.calorie ) {
         if ( updateUserDTO.password || updateUserDTO.access ) {
           return ServiceResponse.error('ask your MANAGER to update your password/access');
@@ -94,14 +100,14 @@ export default class UserController {
         // if ( updateUserDTO.password || updateUserDTO.access ) {
         //   return 'You cannot update password/access Ask MANAGER or ADMIN to do so';
         // }
-        updateUserDTO.userName = await this.accessService.getCurrentUser();
+        updateUserDTO.userName = thisUser.userName;
         return ServiceResponse.success(await this.userService.updateUserExpectation({
           userName: updateUserDTO.userName,
           calorie: updateUserDTO.calorie,
         }));
       }
-    } else if (await this.accessService.getAccess() === EAccess.MANAGER || await this.accessService.getAccess() === EAccess.ADMIN) {
-      const user: UserInterface = await this.userService.findByUserName(updateUserDTO.userName);
+    } else if (thisUser.access === EAccess.MANAGER || thisUser.access === EAccess.ADMIN) {
+      const user: UserEntity = await this.userService.findByUserName(updateUserDTO.userName);
       if (!user) {
         return ServiceResponse.error(`user \"${updateUserDTO.userName}\" not found`);
       } else {
@@ -126,7 +132,7 @@ export default class UserController {
             return ServiceResponse.error('Only one value can be updated at a time');
           }
         } else {
-          if ( await this.accessService.getAccess() === EAccess.ADMIN ) {
+          if ( thisUser.access === EAccess.ADMIN ) {
             if (updateUserDTO.password || updateUserDTO.access) {
               return ServiceResponse.error('Only one value can be updated at a time');
             } else {
@@ -138,8 +144,6 @@ export default class UserController {
           }
         }
       }
-    } else {
-      return ServiceResponse.error('Please LOG IN to perform an action');
     }
   }
 }
